Fix the rest of the Sphinx documentation errors.
This goes through and fixes all remaining documentation warnings for
syntax errors, broken references, conflicting types, files missing from
the Table of Contents, and more.

One change that's annoying but necessary to work around a Sphinx bug is
that we had to filter out any functions named "list" on classes. This
is because Sphinx will prioritize references to those instead of falling
back on the intersphinx mapping in the Python docs for the type. This
means our docs are incomplete, but in practice, the only time we name a
function "list" is when subclassing a Django template loader.

Added a build step to the Travis CI rules to ensure our docs are
generating without warnings or errors.

Testing Done:
Docs built without any warnings or errors.

Checked that the "list" references pointed to the Python docs.

Reviewed at https://reviews.reviewboard.org/r/7597/