Prepare the extensions code for a migration away from Django Evolution.
The extensions code has a firm dependency on Django Evolution, but down
the road, we're definitely not going to want that.

When a Django app using ExtensionManager loads, it imports
django_evolution, which on Django 1.8/1.9 results in a deprecation
warning, due to the import occurring too early in the Django
initialization cycle.

To address both of these, the migration code has been split off into its
own function that handles the importing of django_evolution. Later,
we'll be able to make this a bit smarter and have it handle official
Django migrations directly.

Testing Done:
Unit tests pass.

No longer saw the import warnings against Django 1.8.

Reviewed at https://reviews.reviewboard.org/r/7610/