Add a command to install the extension media of an extension
Previously, there was no way to install the media of an extension
outside if its intsall procedure. Now there is a management command,
`install-extension-media`, that will install the extension media of an
extension. This behaviour can be forced (e.g., if the extension is
already installed) with the `--force` option. Either a single
extension's media will be installed (if `--extension-id` is provided),
or the extension media of all extensions will be installed.

Testing Done:
Installed the extension media for a single extension and all enabled
extensions. Verified `--force` forces installation.

Reviewed at https://reviews.reviewboard.org/r/7627/