Allow datagrids to build their own paginators.
DataGrid subclasses can now override build_paginator() in order to
provide their own paginator class, or to process the queryset before
handing it to the paginator.

This is particularly useful when integrating with Haystack, as some
additional work needs to be done to fetch the right items in the
queryset before giving it to a paginator.

Testing Done:
Unit tests pass.

Used this along with some changes in Splat, which were able to override
and build the paginator itself.

Reviewed at https://reviews.reviewboard.org/r/7631/