Prevent unnecessary queries in the datagrid for None values.
The query optimizer in the datagrid code wasn't checking if associated
IDs were None, and still tried to look them up from the database. This
was silly and pointless. It no longer tries to do such things.

Testing Done:
Saw this appear in a list of queries before-hand. Fixed it, and saw that
the queries were no longer appearing.

Reviewed at https://reviews.reviewboard.org/r/7632/