Simplify save_image_to_storage.
The save_image_to_storage template tag is one of the most ancient parts of
Djblets, and it was being way more complicated and fragile than necessary.
Because storage objects now provide an `open` method that returns a `File`
object, we can just write the image directly to storage rather than doing a
temporary file dance.

This also goes through and renames variables to not be called "file".

Testing done:
Used this with the PDF thumbnailer, which was encountering some raciness in my
setup.

Reviewed at https://reviews.reviewboard.org/r/7711/