Add thorough documentation for the djblets_utils template tags.
djblets_utils is one of the oldest modules in Djblets, and its
documentation was really lacking. This change fleshes it out, bringing
it up to current standards, and adds examples showing how to use all the
tags.

This started off as some cleanups to appease Sphinx, but ended up
becoming a complete rewrite of the docs.

Testing Done:
Built the docs and read through them.

Reviewed at https://reviews.reviewboard.org/r/7738/