Fix errors with datagrid sorting with trailing commas in column lists.
Bingbot's been causing some odd problems on our server, attempting to
access URLs with a sort column list that contains a trailing comma. This
meant that we had a blank entry in our parsed list of columns, which
crashed some very old code.

This might be a link we're generating somewhere, but we don't want to
crash in this case. We now ensure that the list from the client only
contains valid values.

Testing Done:
Tested with a trailing comma. The page loaded, instead of crashing.

Reviewed at https://reviews.reviewboard.org/r/7836/