Add a convenience function for getting static file URLs for extensions.
This introduces `Extension.get_static_url()`, which takes a static media
file bundled with the extension and returns the URL for it. This is
equivalent to what the ext_static template tag does, and in fact that
template tag has been updated to use this. It's much easier now to
reference extension-provided static media files outside of templates.

Testing Done:
Tested that templates were still able to reference the static files.

Made use of the new function in some of the new integrations work.

Reviewed at https://reviews.reviewboard.org/r/7854/