Centralize the templates used for administration forms.
We had a number of templates that were piggy-backing on the logic in the
siteconfig/settings.html template, but this wasn't a very good
organizational approach. These templates weren't related to siteconfig
settings, and there was a theoretical possibility that a
siteconfig-related change would break the other templates.

We now have a base template, djblets_forms/admin/change_form_page.html,
which can be relied upon as a base template for any forms that are to
appear in the administration UI. It contains the same logic as the
original form, with the additions of including the js18n library,
defining the jQuery object, and allowing the save message to be
configured.

Other templates have been updated to use this as the new base, and
backwards-compatibility templates have been added.

There's also a fix for the margin at the bottom of the extension form,
which was placed on the wrong element, causing a visual issue.

Testing Done:
Tested the extension configuration, site settings, and upcoming
integration pages. I was able to save each without problems.

Reviewed at https://reviews.reviewboard.org/r/7851/