Show an appropriate error when trying to load an uninstalled extension.
When an extension is uninstalled, we still have the entry in the
database with the configuration for that extension. However, since
it couldn't be loaded, we try to show an error message, but never get
far enough to compute one.

We now have a suitable error in the API in the case where the extension
class doesn't exist, saying that the extension isn't installed. This
will help with this confusing error state.

Testing Done:
Loaded my extension list and saw reasonable error messages, instead of
a blank error.

Reviewed at https://reviews.reviewboard.org/r/7855/