Update PIL import
Previously, if PIL's `Image` library could not be imported, the `Image`
module would be imported. However, this is not a good way to import PIL
becuase it masks the errors from the first import, which can occur if
some dependencies (such as `libtiff` or `libjpeg`) are missing.

Testing Done:
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/7876/