Add a registry for configuration forms
Review Board's "account pages", which are Djblets `ConfigPage`s with
dynamically determined forms have been ported to Djblets. The behaviour
for dynamically configuring a form is encapsulated in the
`DynamicConfigPageMixin` and the `ConfigPageRegistry`. Config pages
that are to be dynamic should inherit from the `DynamicConfigPageMixin`
and are to be managed by a single registry; multiple registries can
manage multiple sets of configuration pages, however.

Testing Done:
- Ran unit tests.
- Used this with an upcoming change for Review Board.

Reviewed at https://reviews.reviewboard.org/r/7827/