Disable coverage reports and enable test IDs for the Djblets test suite.
This turns off coverage reports for Djblets by default, requiring the
--with-coverage parameter to enable them (like we require for RBTools
and Review Board), reducing the test time to 25% of what it was before.

It also enables test IDs, which makes it easy to re-run a failing test
without hunting for the proper test function.

Testing Done:
Ran the test suite. Saw a huge performance increase, with test IDs and
no coverage report. Then ran again with a coverage report, and saw that.

Reviewed at https://reviews.reviewboard.org/r/7892/