Use a more modern style for code blocks in markdown text.
In our markdown style, we showed inline code in blue, and code blocks were
indented. This wasn't great for readability, and the world has moved on. This
change mimics the Slack style for code blocks, which has started to percolate
out into the rest of the world (GitHub also shows code blocks with a grey
background these days).

While I was in here I noticed that a few things used a "pre-wrap" macro which
wasn't defined. I've added a definition for that.

Testing done:
Created a description with an inline code span, a code block with syntax
highlighting, and a code block without. Saw that everything looked very nice.

Reviewed at https://reviews.reviewboard.org/r/7959/