Fix display of codehilite sections which have been iterated.
When using `iter_markdown_lines` on something that contains a codehilite block,
the resulting output would display in a very ugly way (and a way that was not
valid HTML). I've improved the iteration and the style for these so they show
correctly.

Testing done:
- Looked at a change description in Review Board that added some code blocks.
  Before, they looked very ugly, now they look pretty.
- Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/7983/