Add an Avatar Service Registry to Djblets
The avatar services registry is a new object for managing avatar
services, which are defined to provide multiple ways of displaying
users' avatars. Currently, the only supported service is the Gravatar
(https://gravatar.com) service.

Avatars support the `srcset` attribute for `<img>` tags. In browsers
where it is not supported, a polyfill is used.

Testing Done:
- Ran unit tests
- Used this with an upcoming patch in Review Board.
- Tested the `srcset` polyfill in IE11 with Review Board.

Reviewed at https://reviews.reviewboard.org/r/7844/