Fix an error in logging in jquery.gravy.retina.js.
The logging statement when we fail to parse the URL for a gravatar was using an
undefined name, which was showing up as an error in the console, but more
importantly, could break some js-tests that used fake gravatar URLs.

Testing done:
Ran js-tests.

Reviewed at https://reviews.reviewboard.org/r/8029/