Fix a couple small errors in the Retina avatar support.
The new support for Retina avatars was assuming the existence of a
"srcset" attribute on images, and failing with a JavaScript error if not
set. The error then bubbled up, impacted any callers, breaking them in
the process. We now assume a default of an empty srcset, falling back on
just using src.

It also failed to return the current element set, preventing chaining
from working. We now just return 'this' at the end.

Testing Done:
Encountered the errors when using the RB reply UI. This change fixed them.

Reviewed at https://reviews.reviewboard.org/r/8035/