Fix extra item removal in ListItemView.
When an item is being removed because of its destroy signal, we were removing
an extra item as well. What was happening is that the item's view would notice
that its underlying model was destroyed, and it would remove itself. The list
view would then see the removal from the collection, and remove whatever
element was at that same index.

I've changed it so that instead of blindly selecting an element by index, we
instead keep an array of the child views and look up which one needs to be
removed by its model.

Testing done:
Ran js-tests that used this (and were failing before).

Reviewed at https://reviews.reviewboard.org/r/8069/