Fix rendering for config form JS models
The JS model for a JS view for a config form was previously rendering
incorrectly (without a trailing `)`). This issue has been addressed.
Also, we now are more careful to only render a trailing comma after the
JS model data if there is in fact a page-wide config object. This object
is now only added to the models if it exists.

Testing Done:
Tested this with Review Board. The generated JS works now.

Reviewed at https://reviews.reviewboard.org/r/8082/