Add documentation on the new integrations framework.
This includes guides on how to consume the integrations framework in a
project, and how to write new integrations.

The guide on consuming the integrations framework walks developers
through the classes they need to create, settings to set, and other
responsibilities for the consuming application.

The guide on writing integrations covers how to build an integration,
manage configuration, and more.

This also sets up the codebase docs for the integrations module.

Testing Done:
Built and read through the docs, checking for links and typos.

Reviewed at https://reviews.reviewboard.org/r/7948/