Fix TestModelsLoaderMixin with real models.py on Django 1.7+.
TestModelsLoaderMixin wasn't properly handling the case where a real
models.py file was being specified when on Django 1.7+. It assumed a
fake module was set up and tried using it, but in reality, it would be
None, causing failures in tests.

This is a trivial change that adds a conditional around that code.