Add signals and attributes for datagrid mobile mode
Datagrids now represent their current mobile state through the
`data-datagrid-mobile-mode` attribute. When the datagrid changes modes,
it will update that attribute and fire the `mobileModeToggle` signal with
an object that indicates wether it was enabled or disabled.

Testing Done:
Tested this with an upcoming change in Splat.

Reviewed at https://reviews.reviewboard.org/r/8132/