Allow the register view to take extra form keyword arguments
`djblets.auth.views.register` now accepts a `form_kwargs` argument that
specifies a dictionary of keyword arguments to pass to the form class
upon instantiation. This allows more specialization of the form class
that is used and will allow us to avoid the factory pattern when extra
data is required by the form.

Testing Done:
Used this in an upcoming Splat patch.

Reviewed at https://reviews.reviewboard.org/r/8191/