Add convenience methods to the avatar registry
The avatar services registry now provides a convenience method to lookup
an avatar service directly by its avatar service ID (as that is all that
is used), as well as a method to get either the requested service or the
default.

In addition, the `get_defaults()` method has been refactored to reference
a list of avatar service classes on the registry itself to simplify
subclass implementation.

The `enabled_services` setter now also computes the sets of avatars
services to disable and enable and disables (or enables) each one
individually.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8203/