Fix configuration forms not rendering JS config properly
Previously, there was an issue of the config forms having incorrectly
rendered JS. That issue has been fixed, but there was also an issue of
the wrong variable names being splatted and so the generated config was
syntactically correct, albeit empty. This patch corrects that.

This patch also adds some closing tags that were missing from the
template.

Testing Done:
Used this with an upcoming change in Splat.

Reviewed at https://reviews.reviewboard.org/r/8202/