Only animate the datagrid menu when it is not animating
Previously, toggling the column menu for datagrids fast enough would
result in it breaking due to a variable being `null`. Now, we will only
trigger the menu opening/closing animation if the menu is not currently
being animated. The code has also been cleaned up a bit to use a flag
instead of a jQuery object to determine whether or not the menu is open.

Testing Done:
Could not reproduce this issue with the patch applied.

Reviewed at https://reviews.reviewboard.org/r/8198/