Allow integration forms to use admin widgets and the current HttpRequest.
The integration configuration forms now allow the use of admin widgets,
like the side-by-side multi-select field. The necessary JavaScript files
are loaded onto the page to support those widgets.

The configuration forms also now take a request parameter, which allows
subclasses to provide options based on the request (such as restricting
objects for the configuration for the owning user or organization).

Testing Done:
Made use of the request and admin widgets in an integration.

Reviewed at https://reviews.reviewboard.org/r/8167/