Add customizable avatar services
Djblets now supports avatar services that can be customized per-user via
a configuration form. It currently includes one such avatar service, the
`FileUploadAvatarService`, which allows users to upload an avatar of
their choosing.

The customization form uses nested configuration page forms to
dynamically show and hide the required fields, as well as include the
necessary CSS and JavaScript bundles for the forms.

Each avatar service that is customizable requires a `SettingsManager`,
which is set per-registry. The settings manager is responsible for
loading and saving the per-user configuration of each avatar service.
Djblets does not come with a default one (as it depends on the database
schema).

Testing Done:
- Ran unit tests.
- Changed my avatar to an uploaded avatar in Review Board.

Reviewed at https://reviews.reviewboard.org/r/8205/