Add field serialization/deserialization support to KeyValueForm.
This adds support for defining serializers/deserializers for fields in
KeyValueForm, allowing a form field that returns a non-JSON-serializable
object to be stored serialized, and converting it back to the object on
load.

This is most helpful when using a complex field that returns objects
that need special serialization, such as database models.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/8261/