Update jshintrc to support ES6 and not warn about unused parameters.
This gets jshint set up to support validating ES6, for some upcoming
work happening in Djblets. It also tweaks the unused variable option to
only warn about actual variables and not function parameters (which we
often want to declare but not use, such as in a base class).

Testing Done:
Ran jshint against some upcoming code. Saw that it no longer complained
about the unused function parameters in a base class, and that it had no
problem with the new ES6 code.

Reviewed at https://reviews.reviewboard.org/r/8277/