Change reference of form.error_dict to form.errors
Django forms do not have an `error_dict` attribute. The correct
attribute is the `errors` attribute. This was causing form-wide errors
not to appear in some cases.

Testing Done:
Submitted a form with errors. Saw the error message at the top of the
form.

Reviewed at https://reviews.reviewboard.org/r/8284/