Update Djblets doc building for beanbag-docutils and new Django 1.6 docs.
This updates our Djblets docs building to use beanbag-docutils instead
of bundling a bunch of extensions, cleaning up some of the tree.

It also switches us over to a fork of the Django 1.6 docs, hosted on
readthedocs.org. The official Django website no longer hosts docs for
unsupported versions of Django, so we're switching to this fork in order
to ensure that we're still linking to appropriate docs.

Testing Done:
Built the docs successfully.

Tested that references to Django docs were working.

Reviewed at https://reviews.reviewboard.org/r/8292/