Fix redundant test runs when specifying specific tests.
The test runner takes a list of test labels (module/class/function
paths) to run, which we were appending to the nose argument list.
However, in our standard cases, these may already be in the nose
argument list, leading to the tests being run more than once.

We still want to respect that argument, so we're now adding only the
entries not otherwise found in the argument list.

Testing Done:
Ran a subset of Review Board's unit tests and saw that the tests were
only run once, not twice.

Reviewed at https://reviews.reviewboard.org/r/8323/